extract data into thier own files
