+++ 01-starting-project/src/app/tasks/task/task.component.ts
@@ -1,12 +1,5 @@
 import { Component, Input } from '@angular/core';
-
-interface Task {
-  id: string;
-  userId: string;
-  title: string;
-  summary: string;
-  dueDate: string;
-}
+import { Task } from './Tasks.model';
 
 @Component({
   selector: 'app-task',
