+++ 01-starting-project/src/app/user/users.model.ts
@@ -1,0 +1,5 @@
+export interface User {
+    id: string;
+    avatar: string;
+    name: string;
+  }
